Refactor the code to reuse it. Add logic to insert data to temporal table correctly
