+++ internal/action/root.go
@@ -1,13 +1,23 @@
-package common
+package action
 
 import (
+	"cloud.google.com/go/bigquery"
 	"context"
 	"log"
 	"time"
-	"cloud.google.com/go/bigquery"
 )
 
-func CheckBigQueryJob(job *bigquery.Job, err error) {
+var bigQueryClient *bigquery.Client
+
+func createBigQueryClient(ctx context.Context, projectId string) *bigquery.Client {
+	client, err := bigquery.NewClient(ctx, projectId)
+	if err != nil {
+		log.Fatalf("→ BQ →→ bigquery.NewClient: %v", err)
+	}
+	return client
+}
+
+func checkBigQueryJob(job *bigquery.Job, err error) {
 	if err != nil {
 		log.Fatal("→ BQ →→ Error creating job", err)
 	}
